+++ src/components/ImpactCategories.tsx
@@ -2,7 +2,7 @@ import {
   type ImpactCategory,
   impactCategoryLabels,
 } from "~/hooks/useCategories";
-import { Tag } from "./Tag";
+import { Tag } from "./ui/Tag";
 
 export const ImpactCategories = ({ tags }: { tags: ImpactCategory[] }) => (
   <div className="flex gap-1">
